Defined methods for dummy classes. Defined missing variables.

+++ c_cpp/SLAM/basedef.h
@@ -12,10 +12,10 @@
 #include <cmath>
 #include <list>
 
-#define PI (std::atan(1) * 4)
-#define rtod(deg)  ((deg) * PI / 180)
-#define dtor(rad)  ((rad) * 180 / PI)
-#define sign(number) (number >= 0)
+#define PI 				(double) (std::atan(1) * 4)
+#define rtod(deg)  		(double) ((deg) * PI / 180)
+#define dtor(rad)  		(double) ((rad) * 180 / PI)
+#define sign(number) 	(number >= 0)
